Fix camera finalization when there are multiple of the same camera
When you have multiple camera objects referencing the same hardware,
created either deliberately or from running a program multiple times
without a full GC, the finalization of one camera can break
interactions with the other cameras. This results in users of the other
cameras getting the error SPINNAKER_ERR_NOT_INITIALIZED(-1002).

This commit introduces a deferred finalization method based on FFTW.jl,
in combination with a method to avoid releasing a camera which is
currently in use by another consumer.